requirements.txt
pyautogui
Pillow
img2pdf